requirements.txt
matplotlib==3.10.7
numpy==2.3.5
pandas==2.3.3
PyYAML==6.0.3
scikit_learn==1.8.0
seaborn==0.13.2
